refactor: skills ディレクトリ走査の重複実装を skill-loader に集約する
generate-skill-manifest の findSkillDirs と validate-agent-skills の
listSkillPackages は「SKILL.md を持つディレクトリを再帰探索し、SKILL.md
発見時は降下を止める」という同一の下層 walk を各自で再実装していた。この
共有下層を skill-loader の listSkillPackageDirs（includeRoot オプション付き）
へ抽出し、両呼び出し元を委譲に置換する。

listSkillFiles（拡張子/無視名フィルタ・全降下・legacy skill.yaml 短絡を持つ
別セマンティクス）と extractSkillId（frontmatter 欠落時に basename へ graceful
fallback する drift 検出の堅牢性・parseFrontMatter は throw する）は意図的に
据え置く。ヘルパは walk 順（未ソート）を返し、各呼び出し元が従来どおりのソート
を適用する（SKILL.md パスのソートと親ディレクトリのソートは、兄弟名が接頭辞
関係にあると `-`(0x2D) と `/`(0x2F) の比較で順序が食い違うため非等価）。

外部挙動は byte 不変: skills:manifest 出力（docs/data/skill-manifest.json）・
agent-skills:validate・skills:validate の各出力が変更前後で完全一致。
skill-loader は action bundle に入るため dist を Node 22 で再ビルド。

Related #1473
